requirements.txt
flask>=3.0
flask-socketio>=5.3
flask-limiter>=3.5
yfinance>=0.2
apscheduler>=3.10
gunicorn>=21.2
psycopg2-binary>=2.9
numpy>=1.24
pandas>=2.0
pyjwt>=2.8
marshmallow>=3.20
psutil>=5.9
python-dotenv>=1.0
